# Improve trace-time performance of jnp.isscalar