This avoids calling `device_put` on non-jax values within `isscalar`. This should improve trace time for `isscalar` and functions that use it, such as array indexing.

Before:
```python
In [1]: import jax
In [2]: %timeit jax.numpy.isscalar(1)
23.9 μs ± 162 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```
After:
```python
In [1]: import jax
In [2]: %timeit jax.numpy.isscalar(1)
108 ns ± 0.59 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
```
(This is part of addressing #25109)